Add a FreeformScope and drastically rework Freeform composable
