+++ app/src/main/java/ch/heigvd/ihm/stickies/App.kt
@@ -1,9 +1,9 @@
 package ch.heigvd.ihm.stickies
 
 import androidx.compose.runtime.Composable
-import ch.heigvd.ihm.stickies.ui.freeform.Freeform
+import ch.heigvd.ihm.stickies.ui.freeform.Pane
 
 @Composable
 fun App() {
-    Freeform()
+    Pane()
 }
